+++ src/decoder.rs
@@ -1,10 +1,15 @@
-use crate::utils::{cstr_to_str, str_to_cstring};
-use crate::{ChromaDownsamplingAlgorithm, ChromaUpsamplingAlgorithm, HeifError};
-use libheif_sys as lh;
 use std::ffi::CString;
 use std::fmt::{Debug, Formatter};
 use std::ptr;
 use std::sync::Mutex;
+
+use libheif_sys as lh;
+
+use crate::utils::{cstr_to_str, str_to_cstring};
+use crate::{
+    AlphaCompositionMode, ChromaDownsamplingAlgorithm, ChromaUpsamplingAlgorithm, ColorProfileNCLX,
+    HeifError,
+};
 
 static DECODER_MUTEX: Mutex<()> = Mutex::new(());
 
@@ -12,6 +17,7 @@ static DECODER_MUTEX: Mutex<()> = Mutex::new(());
 pub struct DecodingOptions {
     inner: ptr::NonNull<lh::heif_decoding_options>,
     decoder_id: Option<CString>,
+    output_image_nclx_profile: Option<ColorProfileNCLX>,
 }
 
 impl DecodingOptions {
@@ -20,12 +26,20 @@ impl DecodingOptions {
         ptr::NonNull::new(inner_ptr).map(|inner| Self {
             inner,
             decoder_id: None,
+            output_image_nclx_profile: None,
         })
     }
 }
 
 impl Drop for DecodingOptions {
     fn drop(&mut self) {
+        let inner_mut = self.inner_mut();
+        if !inner_mut.color_conversion_options_ext.is_null() {
+            unsafe {
+                lh::heif_color_conversion_options_ext_free(inner_mut.color_conversion_options_ext)
+            };
+            inner_mut.color_conversion_options_ext = ptr::null_mut();
+        }
         unsafe {
             lh::heif_decoding_options_free(self.inner.as_ptr());
         }
@@ -101,29 +115,90 @@ impl DecodingOptions {
     }
 
     pub fn color_conversion_options(&self) -> ColorConversionOptions {
-        let lh_options = self.inner_ref().color_conversion_options;
-        ColorConversionOptions {
-            preferred_chroma_downsampling_algorithm: ChromaDownsamplingAlgorithm::n(
-                lh_options.preferred_chroma_downsampling_algorithm,
-            )
-            .unwrap_or(ChromaDownsamplingAlgorithm::Average),
-            preferred_chroma_upsampling_algorithm: ChromaUpsamplingAlgorithm::n(
-                lh_options.preferred_chroma_upsampling_algorithm,
-            )
-            .unwrap_or(ChromaUpsamplingAlgorithm::Bilinear),
-            only_use_preferred_chroma_algorithm: lh_options.only_use_preferred_chroma_algorithm
-                != 0,
-        }
+        let inner = self.inner_ref();
+        ColorConversionOptions::from_cc_options(&inner.color_conversion_options)
     }
 
     pub fn set_color_conversion_options(&mut self, options: ColorConversionOptions) {
-        let lh_options = &mut self.inner_mut().color_conversion_options;
-        lh_options.preferred_chroma_downsampling_algorithm =
-            options.preferred_chroma_downsampling_algorithm as _;
-        lh_options.preferred_chroma_upsampling_algorithm =
-            options.preferred_chroma_upsampling_algorithm as _;
-        lh_options.only_use_preferred_chroma_algorithm =
-            options.only_use_preferred_chroma_algorithm as _;
+        let inner = self.inner_mut();
+        options.fill_cc_options(&mut inner.color_conversion_options);
+    }
+
+    #[cfg(feature = "v1_20")]
+    pub fn alpha_composition_mode(&self) -> AlphaCompositionMode {
+        let inner = self.inner_ref();
+        AlphaCompositionMode::from_libheif(inner.color_conversion_options_ext)
+    }
+
+    #[cfg(feature = "v1_20")]
+    pub fn set_alpha_composition_mode(&mut self, v: AlphaCompositionMode) {
+        let inner = self.inner_mut();
+        if inner.color_conversion_options_ext.is_null() {
+            inner.color_conversion_options_ext =
+                unsafe { lh::heif_color_conversion_options_ext_alloc() };
+        }
+        v.fill_libheif_cc_options_ext(inner.color_conversion_options_ext);
+    }
+
+    #[cfg(feature = "v1_21")]
+    /// If enabled, it will decode the media timeline,
+    /// ignoring the sequence tracks edit-list.
+    pub fn ignore_sequence_edit_list(&self) -> bool {
+        let inner = self.inner_ref();
+        inner.ignore_sequence_editlist != 0
+    }
+
+    #[cfg(feature = "v1_21")]
+    /// If enabled, it will decode the media timeline,
+    /// ignoring the sequence tracks edit-list.
+    pub fn set_ignore_sequence_edit_list(&mut self, v: bool) {
+        let inner = self.inner_mut();
+        inner.ignore_sequence_editlist = v as _;
+    }
+
+    #[cfg(feature = "v1_21")]
+    pub fn output_image_nclx_profile(&self) -> Option<&ColorProfileNCLX> {
+        self.output_image_nclx_profile.as_ref()
+    }
+
+    #[cfg(feature = "v1_21")]
+    pub fn set_output_image_nclx_profile(&mut self, v: Option<ColorProfileNCLX>) {
+        self.output_image_nclx_profile = v;
+        let profile_ptr = self
+            .output_image_nclx_profile
+            .as_ref()
+            .map(|v| v.inner)
+            .unwrap_or(ptr::null_mut());
+        let inner = self.inner_mut();
+        inner.output_image_nclx_profile = profile_ptr;
+    }
+
+    #[cfg(feature = "v1_21")]
+    /// 0 = let libheif decide (TODO, currently ignored)
+    pub fn num_library_threads(&self) -> u32 {
+        let inner = self.inner_ref();
+        inner.num_library_threads.max(0) as _
+    }
+
+    #[cfg(feature = "v1_21")]
+    /// 0 = let libheif decide (TODO, currently ignored)
+    pub fn set_num_library_threads(&mut self, v: u32) {
+        let inner = self.inner_mut();
+        inner.num_library_threads = v.min(i32::MAX as u32) as _;
+    }
+
+    #[cfg(feature = "v1_21")]
+    /// 0 = use decoder default
+    pub fn num_codec_threads(&self) -> u32 {
+        let inner = self.inner_ref();
+        inner.num_codec_threads.max(0) as _
+    }
+
+    #[cfg(feature = "v1_21")]
+    /// 0 = use decoder default
+    pub fn set_num_codec_threads(&mut self, v: u32) {
+        let inner = self.inner_mut();
+        inner.num_codec_threads = v.min(i32::MAX as u32) as _;
     }
 }
 
@@ -138,7 +213,6 @@ pub(crate) fn get_decoding_options_ptr(
         .unwrap_or_else(ptr::null_mut)
 }
 
-#[repr(C)]
 #[derive(Debug, Copy, Clone)]
 pub struct ColorConversionOptions {
     pub preferred_chroma_downsampling_algorithm: ChromaDownsamplingAlgorithm,
@@ -146,6 +220,50 @@ pub struct ColorConversionOptions {
     /// When set to `false`, libheif may also use a different algorithm
     /// if the preferred one is not available.
     pub only_use_preferred_chroma_algorithm: bool,
+}
+
+impl Default for ColorConversionOptions {
+    fn default() -> Self {
+        let mut cc_options = lh::heif_color_conversion_options {
+            version: 1,
+            preferred_chroma_downsampling_algorithm: 0,
+            preferred_chroma_upsampling_algorithm: 0,
+            only_use_preferred_chroma_algorithm: 0,
+        };
+        unsafe { lh::heif_color_conversion_options_set_defaults(&mut cc_options) };
+        Self::from_cc_options(&cc_options)
+    }
+}
+
+impl ColorConversionOptions {
+    pub fn new() -> Self {
+        Default::default()
+    }
+
+    pub(crate) fn from_cc_options(cc_options: &lh::heif_color_conversion_options) -> Self {
+        let preferred_chroma_downsampling_algorithm =
+            ChromaDownsamplingAlgorithm::n(cc_options.preferred_chroma_downsampling_algorithm)
+                .unwrap_or(ChromaDownsamplingAlgorithm::Average);
+        let preferred_chroma_upsampling_algorithm =
+            ChromaUpsamplingAlgorithm::n(cc_options.preferred_chroma_upsampling_algorithm)
+                .unwrap_or(ChromaUpsamplingAlgorithm::Bilinear);
+        let only_use_preferred_chroma_algorithm =
+            cc_options.only_use_preferred_chroma_algorithm != 0;
+        Self {
+            preferred_chroma_downsampling_algorithm,
+            preferred_chroma_upsampling_algorithm,
+            only_use_preferred_chroma_algorithm,
+        }
+    }
+
+    pub(crate) fn fill_cc_options(&self, cc_options: &mut lh::heif_color_conversion_options) {
+        cc_options.preferred_chroma_downsampling_algorithm =
+            self.preferred_chroma_downsampling_algorithm as _;
+        cc_options.preferred_chroma_upsampling_algorithm =
+            self.preferred_chroma_upsampling_algorithm as _;
+        cc_options.only_use_preferred_chroma_algorithm =
+            self.only_use_preferred_chroma_algorithm as _;
+    }
 }
 
 #[derive(Copy, Clone)]